+++ miracle/data/tests/test_upload.py
@@ -60,12 +60,13 @@ _COMBINED_PAYLOAD['sessions'].extend(_INVALID_SESSIONS)
 
 class DummyTask(object):
 
-    def __init__(self, db=None, stats=None):
+    def __init__(self, bloom_domain=None, db=None, stats=None):
+        self.bloom_domain = bloom_domain
         self.db = db
         self.stats = stats
 
 
-def test_validate():
+def test_validate(bloom_domain):
     url = 'https://example.com/path'
     time = 1469400000
 
@@ -83,9 +84,10 @@ def test_validate():
         {'sessions': [{'url': 'file:///etc/hosts', 'start_time': time}]},
         {'sessions': [{'url': 'https://admin:admin@example.com/',
                        'start_time': time}]},
+        {'sessions': [{'url': 'http://localhost:80/', 'start_time': time}]},
     ]
     for invalid in invalid_inputs:
-        assert not upload.validate(invalid)[0]['sessions']
+        assert not upload.validate(invalid, bloom_domain)[0]['sessions']
 
     valid_inputs = [
         {'sessions': [{'url': url, 'start_time': time, 'duration': None}]},
@@ -94,23 +96,23 @@ def test_validate():
                        'start_time': time, 'duration': None}]},
     ]
     for valid in valid_inputs:
-        assert upload.validate(valid)[0] == valid
+        assert upload.validate(valid, bloom_domain)[0] == valid
 
     corrected_inputs = [(
         {'sessions': [{'url': url, 'start_time': time, 'duration': -100}]},
         {'sessions': [{'url': url, 'start_time': time, 'duration': None}]}
     )]
     for input_, expected in corrected_inputs:
-        assert upload.validate(input_)[0] == expected
+        assert upload.validate(input_, bloom_domain)[0] == expected
 
 
-def test_upload_data_new_user(db, stats):
+def test_upload_data_new_user(bloom_domain, db, stats):
     with db.session(commit=False) as session:
         url = URL(**URL.from_url('http://www.example.com/'))
         session.add(url)
         session.commit()
 
-        task = DummyTask(db=db, stats=stats)
+        task = DummyTask(bloom_domain=bloom_domain, db=db, stats=stats)
         assert upload.upload_data(task, 'foo', _PAYLOAD)
 
         assert session.query(URL).count() == 4
@@ -131,13 +133,13 @@ def test_upload_data_new_user(db, stats):
     ])
 
 
-def test_upload_data_existing_user(db, stats):
+def test_upload_data_existing_user(bloom_domain, db, stats):
     with db.session(commit=False) as session:
         user = User(token='foo')
         session.add(user)
         session.commit()
 
-        task = DummyTask(db=db, stats=stats)
+        task = DummyTask(bloom_domain=bloom_domain, db=db, stats=stats)
         assert upload.upload_data(task, 'foo', _PAYLOAD)
 
         assert session.query(URL).count() == 4
@@ -159,9 +161,9 @@ def test_upload_data_existing_user(db, stats):
     ])
 
 
-def test_upload_data_duplicated_sessions(db, stats):
+def test_upload_data_duplicated_sessions(bloom_domain, db, stats):
     with db.session(commit=False) as session:
-        task = DummyTask(db=db, stats=stats)
+        task = DummyTask(bloom_domain=bloom_domain, db=db, stats=stats)
         assert upload.upload_data(task, 'foo', _PAYLOAD)
         assert upload.upload_data(task, 'foo', _PAYLOAD)
         assert session.query(Session).count() == 10
@@ -173,7 +175,7 @@ def test_upload_data_duplicated_sessions(db, stats):
     ])
 
 
-def test_upload_data_conflict(cleanup_db, db, stats):
+def test_upload_data_conflict(bloom_domain, cleanup_db, db, stats):
     # Use as secondary transaction to insert a conflicting user id,
     # and keep it open while _upload_data runs the first time.
     # Rollback the secondary transaction before _upload_data runs
@@ -195,7 +197,8 @@ def test_upload_data_conflict(cleanup_db, db, stats):
 
             with mock.patch.object(upload, '_upload_data', mock_upload_data):
                 with db.session(commit=False) as session:
-                    task = DummyTask(db=db, stats=stats)
+                    task = DummyTask(bloom_domain=bloom_domain,
+                                     db=db, stats=stats)
                     assert upload.upload_data(
                         task, 'foo', _PAYLOAD,
                         _lock_timeout=100, _retry_wait=0.01)
@@ -204,11 +207,11 @@ def test_upload_data_conflict(cleanup_db, db, stats):
                     assert session.query(Session).count() == 5
 
 
-def test_upload_main(db, stats):
+def test_upload_main(bloom_domain, db, stats):
     def _upload(task, user, data):
         return (user, data)
 
-    task = DummyTask(db=db, stats=stats)
+    task = DummyTask(bloom_domain=bloom_domain, db=db, stats=stats)
     result = upload.main(
         task, 'foo', json.dumps(_PAYLOAD), _upload_data=_upload)
     assert result == ('foo', _PAYLOAD)