+++ miracle/tests/test_bloom.py
@@ -1,0 +1,5 @@
+
+
+def test_domain(bloom_domain):
+    assert 'localhost' in bloom_domain
+    assert 'somefancydomain.com' not in bloom_domain