+++ miracle/worker/app.py
@@ -6,6 +6,7 @@ from celery.signals import (
 )
 from kombu import Queue
 
+from miracle.bloom import create_bloom_domain
 from miracle.bucket import create_bucket
 from miracle.cache import create_cache
 from miracle.db import create_db
@@ -34,9 +35,11 @@ def configure_celery(celery_app):
 
 
 def init_worker(celery_app,
-                _bucket=None, _cache=None, _db=None, _raven=None, _stats=None):
+                _bloom_domain=None, _bucket=None, _cache=None,
+                _db=None, _raven=None, _stats=None):
     configure_logging()
 
+    celery_app.bloom_domain = create_bloom_domain(_bloom=_bloom_domain)
     celery_app.bucket = create_bucket(_bucket=_bucket)
     celery_app.cache = create_cache(_cache=_cache)
     celery_app.db = create_db(_db=_db)
@@ -49,6 +52,8 @@ def init_worker(celery_app,
 
 
 def shutdown_worker(celery_app):
+    celery_app.bloom_domain.close()
+    del celery_app.bloom_domain
     del celery_app.bucket
     celery_app.cache.close()
     del celery_app.cache